refactor: refactor code so that it does not clutter app.js
