+++ src/app.js
@@ -1,8 +1,8 @@
 import Fastify from 'fastify';
 import ejs from 'ejs';
 import fastifyView from '@fastify/view';
-import { getBlogs, saveBlogs } from './utils/db/index.js';
-import { v4 } from 'uuid';
+import { createBlogPage } from './services/blog-page-services/create-blog-page.js';
+import { general } from './services/general/index.js';
 
 const prefix = '/api';
 
@@ -15,36 +15,10 @@ export async function build () {
     }
   });
 
-  fastify.get(prefix, async (request, reply) => {
-    return reply.view('/views/index.ejs', { text: 'text' });
-  });
+  fastify.get(prefix, general);
 
   // This would be the route for create a blog post
-  fastify.post(`${prefix}/blog`, async (request, reply) => {
-    const { body } = request;
-    const { title, author, content} = body;
-    const db = await getBlogs();
-
-    const id = v4();
-
-    const blog = {
-      title,
-      author,
-      content,
-      blogComments: [],
-      createdDate: new Date().getTime(),
-      updatedDate: new Date().getTime()
-    };
-
-    db.blogList[id] = blog;
-
-    await saveBlogs(db);
-
-    return {
-      id,
-      ...blog
-    };
-  });
+  fastify.post(`${prefix}/blog`, createBlogPage);
 
   return fastify;
 }
