+++ src/services/general/index.js
@@ -1,0 +1,3 @@
+export const general = async (request,reply) => {
+    return reply.view('/views/index.ejs', { text: 'text' });
+};
